+++ libs/text/GP_Font.c
@@ -51,7 +51,7 @@ GP_GlyphBitmap *GP_GetGlyphBitmap(const GP_FontFace *font, int c)
 		return NULL;
 	}
 
-	uint16_t offset;
+	uint32_t offset;
 
 	if (font->glyph_offsets[0] == 0)	
 		offset = font->glyph_offsets[i];
